+++ R/read_functions.R
@@ -62,18 +62,22 @@ read_folder <- function(folder, read_function, .clean_file_names = TRUE,
 #'
 col_rep <- function(str, spec_type = "readr") {
   # need spec_type = base, readr, readxl, sql
-  str <- tolower(str)
 
-  str_loc <- stringr::str_locate_all(str, "[a-z\\?\\-][0-9]+")[[1]]
-  str_val <- stringr::str_sub(str, str_loc[,1], str_loc[,2])
+  str_loc <- stringr::str_locate_all(str, "[A-z\\?\\-][0-9]+")[[1]]
 
-  x_exp <- sapply(mapply(rep,
-                         x = stringr::str_extract(str_val, "[a-z\\?\\-]"),
-                         times = stringr::str_extract(str_val, "[0-9]+"),
-                         SIMPLIFY = FALSE),
-                  paste, collapse = "")
+  if (nrow(str_loc) > 0) {
+    str_val <- stringr::str_sub(str, str_loc[,1], str_loc[,2])
 
-  expanded <- stringi::stri_sub_replace_all(str, str_loc[,1], str_loc[,2], replacement = x_exp)
+    x_exp <- sapply(mapply(rep,
+                           x = stringr::str_extract(str_val, "[A-z\\?\\-]"),
+                           times = stringr::str_extract(str_val, "[0-9]+"),
+                           SIMPLIFY = FALSE),
+                    paste, collapse = "")
+
+    expanded <- stringi::stri_sub_replace_all(str, str_loc[,1], str_loc[,2], replacement = x_exp)
+  } else {
+    expanded <- str
+  }
 
   if (spec_type == "readr") return(expanded)
 
@@ -85,4 +89,90 @@ col_rep <- function(str, spec_type = "readr") {
 
   chars <- unlist(strsplit(expanded, split = ""))
   unname(type_map[[spec_type]][chars])
+}
+
+
+#' List Excel tables in a workbook
+#'
+#' \code{get_excel_tables()} is similar to \code{\link[openxlsx]{getTables}()} but
+#' returns more information.
+#'
+#' @inheritParams openxlsx::getTables
+#' @param sheets Character vector of sheet names to list tables from. If `NULL`,
+#' lists all sheets.
+#'
+#' @return A tibble with the worksheet, table name, and cell references.
+#' @export
+#'
+#' @family Read Excel tables
+#'
+#' @examples
+#' example_file <- system.file("examples/excel examples.xlsx", package = "costmisc")
+#'
+#' wb <- openxlsx::loadWorkbook(example_file)
+#' get_excel_tables(wb)
+#'
+get_excel_tables <- function(wb, sheets = NULL) {
+
+  if (!requireNamespace("openxlsx", quietly = TRUE))
+    stop("Package \"openxlsx\" needed for this function to work. Please install it.")
+
+  if (is.null(sheets))
+    sheets <- names(wb)
+
+  sheets %>%
+    purrr::map_df( ~ {
+      tbls <- openxlsx::getTables(wb, .)
+      tibble::tibble(sheet = .,
+                     table = strip_attributes(tbls, FALSE),
+                     range = attr(tbls, "refs"))
+    })
+}
+
+#' Read Excel tables
+#'
+#' \code{read_excel_table()} reads a table object from an Excel workbook.
+#'
+#' @inheritParams openxlsx::getTables
+#' @param table_name A character name of a table to read.
+#' @param table_df A dataframe of table references, built using
+#' \code{\link{get_excel_tables}()}. If `NULL` the table will be built.
+#'
+#' @return A tibble with the table.
+#' @export
+#'
+#' @family Read Excel tables
+#'
+#' @examples
+#' example_file <- system.file("examples/excel examples.xlsx", package = "costmisc")
+#'
+#' wb <- openxlsx::loadWorkbook(example_file)
+#' head(read_excel_table(wb, "tbl_mtcars"))
+#'
+read_excel_table <- function(wb, table_name, table_df = NULL) {
+
+  if (!requireNamespace("openxlsx", quietly = TRUE))
+    stop("Package \"openxlsx\" needed for this function to work. Please install it.")
+
+  if (is.null(table_df))
+    table_df <- get_excel_tables(wb)
+
+  range <- table_df %>%
+    dplyr::filter(table == table_name) %>%
+    dplyr::select(sheet, range)
+
+  if (nrow(range) != 1) stop(paste0("Table \"", table_name, "\" not uniquely found"))
+
+  parse_range <- function(range) {
+    cells <- stringr::str_split(range, ":", simplify = T)
+    colrange <- openxlsx::convertFromExcelRef(cells)
+    rowrange <- stringr::str_remove(cells, "[^0-9]")
+
+    list(cols = colrange[1]:colrange[2],
+         rows = rowrange[1]:rowrange[2])
+  }
+
+  read_cells <- parse_range(range[[2]])
+
+  openxlsx::readWorkbook(wb, sheet = range[[1]], rows = read_cells$rows, cols = read_cells$cols)
 }