+++ NEWS.md
@@ -1,3 +1,16 @@
+# costmisc 0.4.2
+
+* Tweaked `insensitive()` so that it passes through all the join arguments.
+
+# costmisc 0.4.1
+
+* Added `add_missing_column()` to add a column if missing to a data frame.
+
+# costmisc 0.4.0
+
+* Added `get_excel_tables()` to view the Tables from an Excel workbook.
+* Added `read_excel_table()` to read a Table from an Excel workbook.
+
 # costmisc 0.3.0
 
 * Moved in `clean_by()`, `paste_difftime()`, `insensitive()` and `distinct_insensitive()` from `ff2db`.